+++ .github/workflows/test.yml
@@ -10,10 +10,10 @@ env:
 jobs:
   test:
     name: Test with ${{ matrix.go }} and CROSS_COMPILE=${{ matrix.cross_compile }}
-    runs-on: ubuntu-18.04
+    runs-on: ubuntu-latest
     strategy:
       matrix:
-        go: [1.16, 1.17]
+        go: [1.18, 1.19]
         cross_compile: [true, false]
     steps:
       - uses: actions/checkout@v2
@@ -27,13 +27,13 @@ jobs:
         run: make cross
   docs:
     name: Build docs and Coveralls integration
-    runs-on: ubuntu-18.04
+    runs-on: ubuntu-latest
     needs: test
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-go@v2
         with:
-          go-version: 1.16
+          go-version: 1.19
       - name: Create .coverprofile for each targeted directory by re:running tests
         run: make test
       - name: Collect all .coverprofile files and save it to one file gover.coverprofile
