+++ go.mod
@@ -4,8 +4,6 @@ go 1.13
 
 replace github.com/Sirupsen/logrus => github.com/sirupsen/logrus v1.8.1
 
-replace github.com/docker/libcompose => github.com/docker/libcompose v0.4.1-0.20190808084053-143e0f3f1ab9
-
 replace github.com/docker/cli => github.com/docker/cli v20.10.16+incompatible
 
 replace github.com/xeipuuv/gojsonschema => github.com/xeipuuv/gojsonschema v1.2.1-0.20201027075954-b076d39a02e5
@@ -31,28 +29,25 @@ replace github.com/emicklei/go-restful => github.com/emicklei/go-restful v2.16.0
 replace golang.org/x/crypto => golang.org/x/crypto v0.0.0-20220826181053-bd7e27e6170d
 
 require (
+	github.com/compose-spec/compose-go v1.8.2
 	github.com/containerd/containerd v1.6.4 // indirect
 	github.com/deckarep/golang-set v1.7.1
-	github.com/docker/cli v0.0.0-20191017083524-a8ff7f821017
-	github.com/docker/go-connections v0.4.0
-	github.com/docker/libcompose v0.4.0
 	github.com/fatih/structs v1.1.0
 	github.com/fsouza/go-dockerclient v1.6.6
-	github.com/google/go-cmp v0.5.8
+	github.com/google/go-cmp v0.5.9
 	github.com/google/shlex v0.0.0-20191202100458-e7afc7fbc510
 	github.com/joho/godotenv v1.3.0
 	github.com/novln/docker-parser v1.0.0
 	github.com/openshift/api v0.0.0-20200803131051-87466835fcc0
 	github.com/pkg/errors v0.9.1
-	github.com/sirupsen/logrus v1.8.1
+	github.com/sirupsen/logrus v1.9.0
 	github.com/spf13/cast v1.3.1
 	github.com/spf13/cobra v1.1.3
 	github.com/spf13/viper v1.7.1
-	golang.org/x/net v0.0.0-20220906165146-f3363e06e74c // indirect
 	golang.org/x/text v0.3.8 // indirect
 	golang.org/x/tools v0.1.12
 	gopkg.in/yaml.v2 v2.4.0
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
+	gopkg.in/yaml.v3 v3.0.1
 	k8s.io/api v0.22.5
 	k8s.io/apimachinery v0.22.5
 )
