+++ script/test/fixtures/change-in-volume/output-k8s-empty-vols-template.yaml
@@ -59,6 +59,7 @@ spec:
         kompose.service.type: headless
       creationTimestamp: null
       labels:
+        io.kompose.network/change-in-volume-default: "true"
         io.kompose.service: redis
     spec:
       containers:
@@ -67,6 +68,22 @@ spec:
           resources: {}
       restartPolicy: Always
 status: {}
+
+---
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  creationTimestamp: null
+  name: change-in-volume-default
+spec:
+  ingress:
+    - from:
+        - podSelector:
+            matchLabels:
+              io.kompose.network/change-in-volume-default: "true"
+  podSelector:
+    matchLabels:
+      io.kompose.network/change-in-volume-default: "true"
 
 ---
 apiVersion: apps/v1
@@ -87,6 +104,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/change-in-volume-default: "true"
         io.kompose.service: web
     spec:
       containers:
@@ -106,3 +124,4 @@ spec:
         - emptyDir: {}
           name: code-volume
 status: {}
+
