+++ script/test/fixtures/configmap-volume/output-k8s-withlabel.yaml
@@ -21,6 +21,7 @@ spec:
         kompose.volume.type: configMap
       creationTimestamp: null
       labels:
+        io.kompose.network/configmap-volume-default: "true"
         io.kompose.service: db
     spec:
       containers:
@@ -55,6 +56,22 @@ metadata:
   name: db-cm0
 
 ---
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  creationTimestamp: null
+  name: configmap-volume-default
+spec:
+  ingress:
+    - from:
+        - podSelector:
+            matchLabels:
+              io.kompose.network/configmap-volume-default: "true"
+  podSelector:
+    matchLabels:
+      io.kompose.network/configmap-volume-default: "true"
+
+---
 apiVersion: apps/v1
 kind: Deployment
 metadata:
@@ -77,6 +94,7 @@ spec:
         kompose.volume.type: configMap
       creationTimestamp: null
       labels:
+        io.kompose.network/configmap-volume-default: "true"
         io.kompose.service: web
     spec:
       containers:
