+++ script/test/fixtures/configmap-volume/output-os-withlabel.yaml
@@ -19,6 +19,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/configmap-volume-default: "true"
         io.kompose.service: db
     spec:
       containers:
@@ -114,6 +115,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/configmap-volume-default: "true"
         io.kompose.service: web
     spec:
       containers:
