+++ script/test/fixtures/configmap-volume/output-os.yaml
@@ -17,6 +17,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/configmap-volume-default: "true"
         io.kompose.service: db
     spec:
       containers:
@@ -110,6 +111,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/configmap-volume-default: "true"
         io.kompose.service: web
     spec:
       containers:
