+++ script/test/fixtures/deploy/placement/output-placement-os.yaml
@@ -34,6 +34,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/placement-default: "true"
         io.kompose.service: redis
     spec:
       affinity:
