+++ script/test/fixtures/envvars-interpolation/output-k8s.yaml
@@ -16,6 +16,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/envvars-interpolation-default: "true"
         io.kompose.service: myservice
     spec:
       containers:
@@ -32,4 +33,20 @@ spec:
           resources: {}
       restartPolicy: Always
 status: {}
+
+---
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  creationTimestamp: null
+  name: envvars-interpolation-default
+spec:
+  ingress:
+    - from:
+        - podSelector:
+            matchLabels:
+              io.kompose.network/envvars-interpolation-default: "true"
+  podSelector:
+    matchLabels:
+      io.kompose.network/envvars-interpolation-default: "true"
