+++ script/test/fixtures/service-group/output-k8s.yaml
@@ -35,6 +35,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/service-group-default: "true"
         io.kompose.service: librenms-dispatcher
     spec:
       containers:
@@ -81,4 +82,20 @@ spec:
     requests:
       storage: 100Mi
 status: {}
+
+---
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  creationTimestamp: null
+  name: service-group-default
+spec:
+  ingress:
+    - from:
+        - podSelector:
+            matchLabels:
+              io.kompose.network/service-group-default: "true"
+  podSelector:
+    matchLabels:
+      io.kompose.network/service-group-default: "true"
