+++ script/test/fixtures/statefulset/output-k8s.yaml
@@ -56,6 +56,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/statefulset-default: "true"
         io.kompose.service: db
     spec:
       containers:
@@ -95,6 +96,22 @@ status:
   replicas: 0
 
 ---
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  creationTimestamp: null
+  name: statefulset-default
+spec:
+  ingress:
+    - from:
+        - podSelector:
+            matchLabels:
+              io.kompose.network/statefulset-default: "true"
+  podSelector:
+    matchLabels:
+      io.kompose.network/statefulset-default: "true"
+
+---
 apiVersion: apps/v1
 kind: StatefulSet
 metadata:
@@ -112,6 +129,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/statefulset-default: "true"
         io.kompose.service: wordpress
     spec:
       containers:
