+++ script/test/fixtures/v3.0/output-os.yaml
@@ -38,7 +38,7 @@ spec:
       creationTimestamp: null
       labels:
         io.kompose.network/app-network: "true"
-        io.kompose.network/normalized-network: "true"
+        io.kompose.network/v30-normalized-network: "true"
         io.kompose.network/web-network: "true"
         io.kompose.service: foo
     spec:
@@ -114,6 +114,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/v30-default: "true"
         io.kompose.service: redis
     spec:
       containers:
