+++ script/test/fixtures/volume-mounts/windows/output-k8s.yaml
@@ -35,6 +35,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/windows-default: "true"
         io.kompose.service: db
     spec:
       containers:
@@ -68,4 +69,20 @@ spec:
     requests:
       storage: 100Mi
 status: {}
+
+---
+apiVersion: networking.k8s.io/v1
+kind: NetworkPolicy
+metadata:
+  creationTimestamp: null
+  name: windows-default
+spec:
+  ingress:
+    - from:
+        - podSelector:
+            matchLabels:
+              io.kompose.network/windows-default: "true"
+  podSelector:
+    matchLabels:
+      io.kompose.network/windows-default: "true"
