+++ script/test/fixtures/volume-mounts/windows/output-os.yaml
@@ -35,6 +35,7 @@ spec:
     metadata:
       creationTimestamp: null
       labels:
+        io.kompose.network/windows-default: "true"
         io.kompose.service: db
     spec:
       containers:
